feat: accessToken에 따라 페이지 전환 기능 추가
