+++ scripts/modules/redirector.js
@@ -1,0 +1,8 @@
+function passAuth(){
+    const existAccessToken = localStorage.getItem("accessToken")
+    if(existAccessToken){
+        location.assign("./folder.html")
+    }
+}
+
+export default passAuth
