+++ scripts/signin.js
@@ -1,7 +1,9 @@
+import passAuth from "./modules/redirector.js";
+passAuth()
 import {renderErrorMSGNode, removeErrorMSGNode} from "./modules/handleErrorClass.js";
 import validator from "./modules/validateSignForm.js"
 import togglePasswordHidden from "./modules/toggleBtn.js";
-import { postSignIn } from "./fetch/sign.js";
+
 
 const emailInputNode = document.querySelector(".signForm_input.email")
 const passwordInputNode = document.querySelectorAll(".signForm_input.password")[0]
@@ -13,6 +15,26 @@ function handleSubmitSignIn(e){
     const email = e.target.email.value
     const password = e.target.password.value
     const payload = {email, password}
+    const postSignIn = async (payload) => {
+        try {
+            const response = await fetch("https://bootcamp-api.codeit.kr/api/sign-in",{
+                method: "POST",
+                headers: { 'Content-Type': 'application/json'},
+                body: JSON.stringify(payload),
+            });
+          
+            if (!response.ok) {
+                throw new Error('Network response was not ok');
+            }
+            const json = await response.json()
+            const {accessToken} = json.data
+            localStorage.setItem("accessToken",JSON.stringify(accessToken)); 
+            location.assign("./folder.html")
+            return json.data
+        } catch (error) {
+            console.error('There was a problem with the fetch operation:', error);
+        };
+    };
     if (postSignIn(payload)){
         location.assign("./folder.html")
     }
